de-monolith Stage 7a: relocate perception_monitor + kinematics_sim to kirra-core
First half of the adapter full de-monolith: move the two remaining lean modules the
ROS2 adapter consumes from the heavy `kirra-runtime-sdk` into `kirra-core`. Pure move
+ re-export — zero adapter changes, zero SDK-internal changes; every consumer keeps
working through the re-export shims.

Moved to kirra-core (bodies proven code-identical by diff; only the file header and the
intra-crate `crate::gateway::kinematics_contract` -> `crate::kinematics_contract` /
`kirra_core::PerceptionDerateEvent` -> `crate::PerceptionDerateEvent` paths changed):
- `kirra_core::perception_monitor` <- `src/gateway/perception_monitor.rs` (1348 lines):
  the Track-C perception-derate monitor (KIRRA-OCCY-PMON-001) — pure stateless
  kinematic-plausibility / range-support guards + the `SharedPerceptionCap` stdlib
  `Arc<RwLock>` cache and its publisher. No async, no heavy deps.
- `kirra_core::kinematics_sim` <- `src/kinematics_sim.rs` (734 lines): the deterministic
  vehicle kinematic simulator (`VehicleState` / `apply_enforcement` / `run_simulation`).

The two SDK modules become thin `pub use kirra_core::…::*;` re-exports, so every existing
`crate::gateway::perception_monitor::*` / `crate::kinematics_sim::*` (and the
`kirra_runtime_sdk::…` paths) hold — the policy layer, the fabric governor, posture_cache's
`SharedPerceptionCap`, the wcet_gate tests, the verifier service binary, the CARLA client,
and the adapter all unchanged.

kirra-core gains a `tracing` dependency (the lightweight logging facade the relocated
simulator's denial-path debug log uses) — a facade only, no runtime, clear of the
forbidden heavy tree (tokio/axum/rusqlite/reqwest/ML). A pre-existing private-intra-doc
link in the monitor's docs was tidied to a plain code span so kirra-core's rustdoc builds
clean under `-D rustdoc::private_intra_doc_links`.

This sets up Stage 7b (the capstone): a small tokio-OK `kirra-capture` crate for the
shared capture writer, then repoint all the adapter's remaining `kirra_runtime_sdk`
imports to `kirra-core` and DROP the adapter's `kirra-runtime-sdk` dependency entirely.

Verification: root workspace build + tests green (kirra-core 138 tests, incl. the
relocated perception_monitor + kinematics_sim suites; none lost — the SDK lib shed exactly
those); CI-equivalent `cargo clippy --workspace -- -D warnings` clean; kirra-core rustdoc
clean under `-D rustdoc::broken_intra_doc_links -D rustdoc::private_intra_doc_links`; parko
workspace logic green (parko-core 204, parko-kirra 148, parko-ros2 168); no `#[path]`
hazards; both moves proven code-identical by diff.
